format and button name change
