+++ src/components/filterButtons/FilterButton.module.css
@@ -1,5 +1,4 @@
-.todoList {
+.filterButtons {
   display: flex;
-  flex-wrap: wrap;
   gap: 1rem;
 }
