+++ src/components/filterButtons/index.js
@@ -1,0 +1,13 @@
+import React from "react"
+import styles from "./FilterButton.module.css"
+
+function FilterButtons() {
+  return (
+    <div className={styles.filterButtons}>
+      <button>All</button>
+      <button>Incomplete</button>
+      <button>Complete</button>
+    </div>
+  )
+}
+export default FilterButtons
