+++ src/components/header/Header.module.css
@@ -1,12 +1,12 @@
 body {
   background-color: #f8f9ff;
 }
-.buttons{
-    display: flex;
-    justify-content: space-between;
+.buttons {
+  display: flex;
+  justify-content: space-between;
 }
 
-.filterButtons{
-    display: flex;
-    gap: 1rem;
+.filterButtons {
+  display: flex;
+  gap: 1rem;
 }
