--- src/components/header/filterButtons/FilterButton.module.css
@@ -1,4 +1,0 @@
-.filterButtons{
-    display: flex;
-    gap: 1rem;
-}
